Update: create seperate file for Context and context provider
