+++ src/context/GeneralContext.tsx
@@ -1,11 +1,7 @@
-import {
-  createContext,
-} from 'react';
+import { createContext } from 'react';
 
 import type { GeneralContextType } from '../types/GeneralContextType';
 
-
 const GeneralContext = createContext<GeneralContextType | null>(null);
-
 
 export default GeneralContext;
