+++ src/utils/todoUtils.ts
@@ -1,5 +1,4 @@
-import { PriorityType, StatusType } from "../types/Todotypes";
-
+import { PriorityType, StatusType } from '../types/Todotypes';
 
 function formatDate(date: Date) {
   const newDate = new Date(date);
