+++ src/wrapper/ContexWrapper.tsx
@@ -1,15 +1,14 @@
-import type { ReactNode } from "react";
+import type { ReactNode } from 'react';
 
-
-import GeneralContextProvider from "../context/GeneralContextProvider";
-import { TodoContextProvider } from "../features/Todos/context/TodoContext";
+import GeneralContextProvider from '../context/GeneralContextProvider';
+import TodoContextProvider from '../features/Todos/context/TodoContextProvider';
 
 function ContextWrapper({ children }: { children: ReactNode }) {
-    return (
-      <GeneralContextProvider>
-        <TodoContextProvider>{children}</TodoContextProvider>
-      </GeneralContextProvider>
-    );
-  }
+  return (
+    <GeneralContextProvider>
+      <TodoContextProvider>{children}</TodoContextProvider>
+    </GeneralContextProvider>
+  );
+}
 
 export default ContextWrapper;
